Consolidate Header and header-context
