+++ src/templates/document.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import Breadcrumbs from '../components/Breadcrumbs/index';
+import Breadcrumbs from '../components/Breadcrumbs';
 import Contents from '../components/Contents/Contents';
 import InternalPageNav from '../components/InternalPageNav';
 import MainColumn from '../components/MainColumn';
